+++ engine/engine.h
@@ -3,7 +3,14 @@
 #include "render/camera.h"
 #include "render/renderer.h"
 #include "render/renderobject.h"
-#include "window/window_system.h"
+#include "resource_importer/include/ResourceSystem.h"
+
+#include <cstdint>
+
+namespace ChikaEngine::Platform
+{
+    class IWindow;
+}
 
 #include <vector>
 
@@ -25,6 +32,7 @@ namespace ChikaEngine::Engine
       private:
         Platform::IWindow* _window;
         Render::Camera _camera;
+        ChikaEngine::Resource::ResourceSystem _resourceSystem;
 
         // TODO: 加入 level 等 实现 GamePlay
     };